# apis/app-api/payment/sdkwork-payment-app-api.openapi.yaml
{
  "openapi": "3.1.2",
  "info": {
    "title": "SDKWork Payment App API",
    "version": "1.0.0",
    "description": "App/client payment execution and refund contract for commerce.payment.",
    "x-sdkwork-api-authority": "sdkwork-payment-app-api",
    "x-sdkwork-sdk-family": "sdkwork-payment-app-sdk",
    "x-sdkwork-audience": "Authenticated SDKWork payment clients"
  },
  "servers": [{ "url": "http://localhost:18094", "description": "Local payment gateway" }],
...
  "security": [{ "AuthToken": [], "AccessToken": [] }],
  "paths": {
...
    "/app/v3/api/refunds": {
      "get": {
        "tags": ["commerce"], "summary": "List refunds.", "operationId": "refunds.list",
        "parameters": [{ "$ref": "#/components/parameters/Page" }, { "$ref": "#/components/parameters/PageSize" }, { "name": "status", "in": "query", "required": false, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Success", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RefundListResponse" } } } }, "default": { "$ref": "#/components/responses/Problem" } },
        "security": [{ "AuthToken": [], "AccessToken": [] }], "x-sdkwork-auth-mode": "dual-token", "x-sdkwork-forbid-credential-headers": false, "x-sdkwork-owner": "sdkwork-payment", "x-sdkwork-api-authority": "sdkwork-payment-app-api", "x-sdkwork-domain": "commerce", "x-sdkwork-resource": "refunds", "x-sdkwork-request-context": "WebRequestContext", "x-sdkwork-api-surface": "app-api", "x-sdkwork-server-request-id": true, "x-sdkwork-permission": "commerce.refunds.read"
      },
      "post": {
        "tags": ["commerce"], "summary": "Create a refund.", "operationId": "refunds.create",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateRefundCommand" } } } },
        "responses": { "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RefundResponse" } } } }, "default": { "$ref": "#/components/responses/Problem" } },
        "security": [{ "AuthToken": [], "AccessToken": [] }], "x-sdkwork-auth-mode": "dual-token", "x-sdkwork-forbid-credential-headers": false, "x-sdkwork-owner": "sdkwork-payment", "x-sdkwork-api-authority": "sdkwork-payment-app-api", "x-sdkwork-domain": "commerce", "x-sdkwork-resource": "refunds", "x-sdkwork-request-context": "WebRequestContext", "x-sdkwork-api-surface": "app-api", "x-sdkwork-server-request-id": true, "x-sdkwork-permission": "commerce.refunds.create", "x-sdkwork-idempotent": true
      }
    },
    "/app/v3/api/refunds/{refundId}": {
      "get": {
        "tags": ["commerce"], "summary": "Retrieve a refund.", "operationId": "refunds.retrieve",
        "parameters": [{ "name": "refundId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Success", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RefundResponse" } } } }, "default": { "$ref": "#/components/responses/Problem" } },
        "security": [{ "AuthToken": [], "AccessToken": [] }], "x-sdkwork-auth-mode": "dual-token", "x-sdkwork-forbid-credential-headers": false, "x-sdkwork-owner": "sdkwork-payment", "x-sdkwork-api-authority": "sdkwork-payment-app-api", "x-sdkwork-domain": "commerce", "x-sdkwork-resource": "refunds", "x-sdkwork-request-context": "WebRequestContext", "x-sdkwork-api-surface": "app-api", "x-sdkwork-server-request-id": true, "x-sdkwork-permission": "commerce.refunds.read"
      }
    }
  },
  "components": {
    "securitySchemes": {
      "AuthToken": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" },
      "AccessToken": { "type": "apiKey", "in": "header", "name": "Access-Token" }
    },
    "parameters": {
      "Page": { "name": "page", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 1, "default": 1 } },
      "PageSize": { "name": "page_size", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 1, "maximum": 200, "default": 20 } },
...
    },
    "responses": {
      "Problem": { "description": "SDKWork problem detail", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ProblemDetail" } } } }
    },
    "schemas": {
      "SdkWorkApiResponse": { "type": "object", "additionalProperties": false, "required": ["code", "data", "traceId"], "properties": { "code": { "type": "integer", "format": "int32", "const": 0 }, "data": { "type": "object" }, "traceId": { "type": "string", "format": "uuid" } } },
...
      "PageInfo": { "type": "object", "additionalProperties": false, "required": ["mode", "page", "pageSize", "totalItems", "totalPages"], "properties": { "mode": { "type": "string", "const": "offset" }, "page": { "type": "integer", "minimum": 1 }, "pageSize": { "type": "integer", "minimum": 1, "maximum": 200 }, "totalItems": { "type": "string", "pattern": "^[0-9]+$" }, "totalPages": { "type": "integer", "minimum": 0 }, "hasMore": { "type": "boolean" } } },
      "ProblemDetail": { "type": "object", "additionalProperties": true, "required": ["type", "title", "status", "code", "traceId"], "properties": { "type": { "type": "string", "format": "uri-reference" }, "title": { "type": "string" }, "status": { "type": "integer", "minimum": 400, "maximum": 599 }, "detail": { "type": "string" }, "code": { "type": "integer", "format": "int32", "minimum": 40001, "maximum": 79999 }, "traceId": { "type": "string", "format": "uuid" } } },
...
      "CreateRefundCommand": { "type": "object", "additionalProperties": false, "required": ["orderId"], "properties": { "orderId": { "type": "string" }, "paymentAttemptId": { "type": "string" }, "amount": { "type": "string", "pattern": "^[0-9]+(\\.[0-9]{1,2})?$" }, "reasonCode": { "type": "string" } } },
...
      "Refund": { "type": "object", "additionalProperties": false, "required": ["refundId", "refundNo", "orderId", "amount", "currencyCode", "status"], "properties": { "refundId": { "type": "string" }, "refundNo": { "type": "string" }, "orderId": { "type": "string" }, "paymentAttemptId": { "type": "string" }, "amount": { "type": "string" }, "currencyCode": { "type": "string" }, "status": { "type": "string" }, "reasonCode": { "type": "string" } } },
...
      "RefundResponse": { "allOf": [{ "$ref": "#/components/schemas/SdkWorkApiResponse" }, { "type": "object", "required": ["data"], "properties": { "data": { "type": "object", "required": ["item"], "properties": { "item": { "$ref": "#/components/schemas/Refund" } } } } }] },
...
      "RefundListResponse": { "allOf": [{ "$ref": "#/components/schemas/SdkWorkApiResponse" }, { "type": "object", "required": ["data"], "properties": { "data": { "type": "object", "required": ["items", "pageInfo"], "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/Refund" } }, "pageInfo": { "$ref": "#/components/schemas/PageInfo" } } } } }] }
    }
  }
}